Inline property-based suite helper
